Moving all Android code that belongs to the library from the example project to the library itself + also use the permissionhandler plugin
